Publish a fork under our coordinates
We need the changes, and our PR on the original project did not receive
an answer. Given the current situation (COVID-19 lockdowns everywhere),
this is totally reasonable. To move forward with our own projects,
however, we decided to roll our own fork to Clojars.

As things need to stay there forever as soon as some 3rd party picks it
up, we can and should do things right here. So I also made some changes
to the README.